Marketplace page tab function completed
