+++ src/App.scss
@@ -202,7 +202,7 @@ span.ant-pagination-item-ellipsis {
         line-height  : 27px;
         border-bottom: 1px solid transparent;
 
-        &:hover {
+        &:hover, &.active {
           border-bottom: 1px solid white;
         }
       }
@@ -261,6 +261,7 @@ span.ant-pagination-item-ellipsis {
         background     : #FFFFFF;
         border-radius  : 7px;
         list-style-type: none;
+        z-index        : 999;
 
         &>.dropdown-item {
           width  : 100%;
