+++ src/components/Chains/style.scss
@@ -28,6 +28,7 @@
         border-radius  : 7px;
         list-style-type: none;
         overflow       : hidden;
+        z-index        : 999;
 
         &>.dropdown-item {
             width      : 100%;
