+++ src/components/Sale/NFT.jsx
@@ -8,7 +8,6 @@ import { ShoppingCartOutlined } from "@ant-design/icons";
 import NFTMeta from "../NFTMeta/NFTMeta";
 import { decodeNftUriToJson } from "../../helpers/nft";
 import { openNotification } from "helpers/notifications";
-import NFTCard from "../../components/NFTCard/NFTCard";
 
 export default function NFT({ tokenId, canMint, mintedTokens, pricePerToken }) {
   const { Moralis } = useMoralis();
@@ -18,7 +17,7 @@ export default function NFT({ tokenId, canMint, mintedTokens, pricePerToken }) {
   const { chainId } = useMoralisDapp();
   const abi = abis.heh_nft;
   const contractAddress = getHoldemHeroesAddress(chainId);
-
+  
   const options = {
     abi,
     contractAddress,
@@ -136,6 +135,5 @@ export default function NFT({ tokenId, canMint, mintedTokens, pricePerToken }) {
     >
       <NFTMeta metadata={nft} />
     </Card>
-    // <NFTCard key={`token_${tokenId}`} nft={nftData} chainId={chainId} nftOwner={nftOwner} canMint={canMint} action={() => postRevealMint(tokenId)} />
   );
 }
